recommender-back/RecommenderServer/requirements.txt
Django==2.0.6
djangorestframework==3.8.2
django-cors-headers==2.3.0